Refactor msgspec backend module into package.
