+++ litestar/dto/factory/backends/__init__.py
@@ -1,6 +1,6 @@
 from __future__ import annotations
 
 from .abc import AbstractDTOBackend
-from .msgspec import MsgspecDTOBackend
+from .msgspec.backend import MsgspecDTOBackend
 
 __all__ = ("AbstractDTOBackend", "MsgspecDTOBackend")
